feat: mine square added, + action and propagation
- Class square created, that represent a part of
the minefield. On click it send the event to the
app that control it. Then, the app reveal the
square content, a mine, a number of mine around or
nothing.
- GUI now display the field in the squareMesh
- GUI fixed : the random level (at the start of
app) was not displayed properly. Manual change
interact with ItemListener so had to add a boolean
to deactivate it temporary.
- App now control the squareMesh, generation,
field initialisation on the first click, and
manage action on square click. Win and loose phase
not completly implemented yet.